+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to the "robotcode" extension will be documented in this file
 - Provide better error messages if python and robot environment not matches RobotCode requirements
   - fixes [#40](https://github.com/d-biehl/robotcode/issues/40)
 - Correct restart of language server client if python interpreter changed
+- Correct start of root test item if `robotcode.robot.paths` is used
 
 ##  0.11.0