+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 All notable changes to the "robotcode" extension will be documented in this file.
 
-## [Unreleased]
+## [Unreleased]
+- none so far
+##  0.11.1
 
 - Provide better error messages if python and robot environment not matches RobotCode requirements
   - fixes [#40](https://github.com/d-biehl/robotcode/issues/40)